refactor(audio): restructure audio processing services
